requirements.txt
requests
json
python-dotenv
mistralai
mistral